+++ src/main/java/org/folio/search/repository/InstanceSubjectRepository.java
@@ -11,6 +11,7 @@ import static org.folio.search.utils.SearchUtils.INSTANCE_SUBJECT_UPSERT_SCRIPT_
 import static org.opensearch.script.ScriptType.STORED;
 
 import java.util.EnumMap;
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
@@ -23,6 +24,7 @@ import org.folio.search.configuration.properties.SearchConfigurationProperties;
 import org.folio.search.domain.dto.FolioIndexOperationResponse;
 import org.folio.search.model.index.SearchDocumentBody;
 import org.folio.search.model.types.IndexActionType;
+import org.folio.search.service.consortium.ConsortiumTenantService;
 import org.opensearch.action.bulk.BulkRequest;
 import org.opensearch.action.update.UpdateRequest;
 import org.opensearch.script.Script;
@@ -33,7 +35,10 @@ import org.springframework.stereotype.Repository;
 @RequiredArgsConstructor
 public class InstanceSubjectRepository extends AbstractResourceRepository {
 
+  private static final String INSTANCE_ID = "instanceId";
+
   private final SearchConfigurationProperties properties;
+  private final ConsortiumTenantService consortiumTenantService;
 
   @Override
   public FolioIndexOperationResponse indexResources(List<SearchDocumentBody> documentBodies) {
@@ -42,9 +47,9 @@ public class InstanceSubjectRepository extends AbstractResourceRepository {
     var docsById = documentBodies.stream().collect(groupingBy(SearchDocumentBody::getId));
     for (var entry : docsById.entrySet()) {
       var documents = entry.getValue();
-      var instanceIds = prepareInstanceIds(documents);
-      if (!IterableUtils.matchesAll(instanceIds.values(), Set::isEmpty)) {
-        var upsertRequest = prepareUpsertRequest(documents.iterator().next(), instanceIds);
+      var instances = prepareInstances(documents);
+      if (!IterableUtils.matchesAll(instances.values(), Set::isEmpty)) {
+        var upsertRequest = prepareUpsertRequest(documents.iterator().next(), instances);
         bulkRequest.add(upsertRequest);
       }
     }
@@ -56,46 +61,53 @@ public class InstanceSubjectRepository extends AbstractResourceRepository {
            : getSuccessIndexOperationResponse();
   }
 
-  private EnumMap<IndexActionType, Set<String>> prepareInstanceIds(List<SearchDocumentBody> documents) {
-    var instanceIds = prepareInstanceIdMap();
+  private EnumMap<IndexActionType, Set<Map<String, Object>>> prepareInstances(List<SearchDocumentBody> documents) {
+    var instances = prepareInstanceMap();
     for (var document : documents) {
       var payload = getPayload(document);
-      var instanceId = String.valueOf(payload.get("instanceId"));
+      var instanceId = String.valueOf(payload.get(INSTANCE_ID));
       if (StringUtils.isNotBlank(instanceId)) {
-        instanceIds.getOrDefault(document.getAction(), instanceIds.get(DELETE)).add(instanceId);
+        var tenantId = document.getTenant();
+        var instance = new HashMap<String, Object>();
+        instance.put(INSTANCE_ID, instanceId);
+        instance.put("tenantId", tenantId);
+        consortiumTenantService.getCentralTenant(tenantId).ifPresent(centralTenant ->
+          instance.put("shared", centralTenant.equals(tenantId)));
+
+        instances.getOrDefault(document.getAction(), instances.get(DELETE)).add(instance);
       } else {
         log.warn("InstanceId is blank in subject event. [payload: {}]", payload);
       }
     }
-    return instanceIds;
+    return instances;
   }
 
-  private EnumMap<IndexActionType, Set<String>> prepareInstanceIdMap() {
-    var instanceIds = new EnumMap<IndexActionType, Set<String>>(IndexActionType.class);
+  private EnumMap<IndexActionType, Set<Map<String, Object>>> prepareInstanceMap() {
+    var instanceIds = new EnumMap<IndexActionType, Set<Map<String, Object>>>(IndexActionType.class);
     instanceIds.put(INDEX, new HashSet<>());
     instanceIds.put(DELETE, new HashSet<>());
     return instanceIds;
   }
 
   private UpdateRequest prepareUpsertRequest(SearchDocumentBody doc,
-                                             EnumMap<IndexActionType, Set<String>> instanceIds) {
+                                             EnumMap<IndexActionType, Set<Map<String, Object>>> instances) {
     return new UpdateRequest()
       .id(doc.getId())
       .scriptedUpsert(true)
       .retryOnConflict(properties.getIndexing().getInstanceSubjects().getRetryAttempts())
       .index(indexNameProvider.getIndexName(doc))
-      .script(new Script(STORED, null, INSTANCE_SUBJECT_UPSERT_SCRIPT_ID, prepareScriptParams(instanceIds)))
-      .upsert(prepareDocumentBody(getPayload(doc), instanceIds), doc.getDataFormat().getXcontentType());
+      .script(new Script(STORED, null, INSTANCE_SUBJECT_UPSERT_SCRIPT_ID, prepareScriptParams(instances)))
+      .upsert(prepareDocumentBody(getPayload(doc), instances), doc.getDataFormat().getXcontentType());
   }
 
-  private Map<String, Object> prepareScriptParams(EnumMap<IndexActionType, Set<String>> instanceIds) {
+  private Map<String, Object> prepareScriptParams(EnumMap<IndexActionType, Set<Map<String, Object>>> instanceIds) {
     return Map.of("ins", instanceIds.get(INDEX), "del", instanceIds.get(DELETE));
   }
 
   private Map<String, Object> prepareDocumentBody(Map<String, Object> payload,
-                                                  Map<IndexActionType, Set<String>> instanceIds) {
-    payload.put("instances", subtract(instanceIds.get(INDEX), instanceIds.get(DELETE)));
-    payload.remove("instanceId");
+                                                  Map<IndexActionType, Set<Map<String, Object>>> instances) {
+    payload.put("instances", subtract(instances.get(INDEX), instances.get(DELETE)));
+    payload.remove(INSTANCE_ID);
     return payload;
   }